Added fire absorbtion
there is an additional area 3d attached to the player head named "FireAbsorber"
the InnerFire node checks this area for both the "FireFloorDetector" Area3D node as well as the "InfiniteFireSource" and "FiniteFireSource" area 3d nodes, and heals the player appropriately in each circumstance, with the Fireball healing the player by its size.
Also very important, the death plane destroys the fires that occasionally get shot through the floor, huge performance save right there.
TLDR to make a prop either a finite or infinte healing pool, just add an area3d node to the root of the prop and name it "InfiniteFireSource" or "FiniteFireSource". Right now the torch is set to Infinite, but you can change that by just renaming the area3d node